bug 936123 - add newsletter signup form to profile
move newsletter subscription form into an include template

show newsletter subscription form on profile edit screen

put profile newsletter subscription behind waffle flag
