+++ apps/landing/templates/landing/apps_newsletter.html
@@ -13,61 +13,7 @@
 
   <section id="content-main" class="full" role="main">
 
-<div class="apps-newsletter-subscribe module">
-<h3 class="mod-title">{{ _('Firefox Apps & Hacks Newsletter') }}</h3>
-{% if form %}
-<form id="newsletter-sub" class="fm-subscribe" action="{{ url('apps_newsletter') }}" method="POST">
-  {{ csrf() }}
-  <p class="mod-intro">{% trans %}
-    Sign up now for news about Firefox OS, Firefox Marketplace and the Open Web apps ecosystem.  <strong>Join us!</strong>
-  {% endtrans %}</p>
-
-  {% if messages %}
-    <ul class="messages">
-      {% for message in messages %}
-        <li {% if message.tags %} class="notice {{ message.tags }}"{% endif %}>{{ message }}</li>
-      {% endfor %}
-    </ul>
-  {% endif %}
-
-    <fieldset>
-      <ol>
-        <li>
-          {{ form.email.errors }}
-          <label for="email">{{ form.email.label }}</label>
-          {{ form.email }}
-        </li>
-        <li>
-          {{ form.country.errors }}
-          <label for="country">{{ form.country.label }}</label>
-          {{ form.country }}
-        </li>
-        <li class="check">
-          {{ form.format.errors }}
-          <span class="label">{{ form.format.label }}</span>
-          <fieldset class="inline">
-            {% for c in form.fields.format.choices %}
-              <label for="sub-{{ c[0] }}"><input type="radio" name="format" value="{{ c[0] }}" id="sub-{{ c[0] }}" checked>{{ c[1] }}</label>
-            {% endfor %}
-          </fieldset>
-        </li>
-        <li class="agree">
-          {{ form.agree.errors }}
-          <label for="id_agree">
-            {{ form.agree }} {{form.agree.label}}
-            {% trans privacy_url='http://www.mozilla.org/en-US/privacy' %}
-              to the <a href="{{ privacy_url }}">Privacy Policy</a>
-            {% endtrans %}
-          </label>
-        </li>
-        <li class="fm-submit"><button type="submit" class="button positive">{{ _('Sign me up!') }}</button></li>
-      </ol>
-    </fieldset>
-    </form>
-{% else %}
-  <p>{{ _('Thank you for signing up for the Apps & Hacks newsletter!') }}</p>
-{% endif %}
-</div>
+  {% include "includes/newsletter_widget.html" %}
 
 </section>
 </div>
